+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
    - Improved file downloading for Kp
    - Added keyword ignore_empty_files to pysat.Instrument and Files objects
     to filter out empty files from the stored file list
+   - Converted all print statements to logging statements
    - Updated cleaning routines for C/NOFS IVM
 - Code Restructure
   - Move `computational_form` to `ssnl`, old version is deprecated